update vehicle entity (+fixture) with the new carDealer entity
